Gpu: Moved creation of Gp0Command to frontend to sync GPU_STAT
GPU_STAT had a sync issue, for example if you set the command GP1(4),
which changes the DMA_DIRECTION flags, there might be some delay for it
to take effect, since the change was happening in the backend thread.
Some games require immediate change or else the game hangs.

Now moved anything that should be synced (creating the commands,
handling any GP1 command that does not need `gpu_context`) in the
frontend.

This improved the accuracy and stability of the GPU unit.
